+++ Controllers/RecipesController.cs
@@ -37,7 +37,7 @@ namespace MEalAPI.Controllers
 
         // GET: api/Recipes/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Recipe>> GetRecipe(int id)
+        public async Task<ActionResult<Recipe>> GetRecipe(long id)
         {
             var recipe = await _context.Recipes
                 .Include(r => r.RecipeIngredients)
@@ -58,7 +58,7 @@ namespace MEalAPI.Controllers
         // PUT: api/Recipes/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutRecipe(int id, Recipe recipe)
+        public async Task<IActionResult> PutRecipe(long id, Recipe recipe)
         {
             if (id != recipe.Id)
             {
@@ -113,7 +113,7 @@ namespace MEalAPI.Controllers
             return NoContent();
         }
 
-        private bool RecipeExists(int id)
+        private bool RecipeExists(long id)
         {
             return _context.Recipes.Any(e => e.Id == id);
         }
